some explicit type casts to float added

+++ readability.c
@@ -14,8 +14,8 @@ int readability(const char *phrase) {
         else if(c == ' ') words++;
         else if (c == '?' || c == '.' || c == '!') sentences++;
 
-    float lettersPer100Words = (letters / words) * 100.0;
-    float sentencesPer100Words = (sentences / words) * 100.0;
+    float lettersPer100Words = ((float)letters / words) * 100;
+    float sentencesPer100Words = ((float)sentences / words) * 100;
 
     float index = (0.0588 * lettersPer100Words) - (0.296 * sentencesPer100Words) - 15.8;
